testing/web-platform/tests/websockets/interfaces/CloseEvent/001.html


<!doctype html>
<title>WebSockets: window.CloseEvent</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
test(function(){assert_not_equals(window.CloseEvent, undefined, 'window.CloseEvent')});
</script>
